+++ R/ondisk.R
@@ -706,7 +706,7 @@ writeHDF5ArrayInImage <- function(
     rhdf5::h5createGroup(h5_path, group = paste0(name, "/spat_", spat))
 
     # write coordinates
-    coords <- vrCoordinates(object, spatial_name = spat)
+    coords <- vrCoordinates(object, spatial = spat)
     if (!inherits(coords, c("DelayedArray", "IterableMatrix")) || replace) {
       if (verbose) {
         message("Writing '", name, "' coordinates")
@@ -721,15 +721,15 @@ writeHDF5ArrayInImage <- function(
         verbose = verbose,
         feature.vs.obs.engine = feature.vs.obs.engine
       )
-      vrCoordinates(object, spatial_name = spat) <- coords
+      vrCoordinates(object, spatial = spat) <- coords
     }
 
     # for each channel
-    channels <- vrImageChannelNames(object, name = spat)
+    channels <- vrImageChannelNames(object, spatial = spat)
     if (!all(grepl("No Channels", channels))) {
       for (ch in channels) {
         # get image and write to h5
-        img <- vrImages(object, name = spat, channel = ch, as.raster = TRUE)
+        img <- vrImages(object, spatial = spat, channel = ch, as.raster = TRUE)
 
         # write image
         if (!inherits(img, "ImageArray") || replace) {
@@ -756,7 +756,7 @@ writeHDF5ArrayInImage <- function(
               verbose = FALSE
             )
           suppressWarnings({
-            vrImages(object, name = spat, channel = ch) <- img
+            vrImages(object, spatial = spat, channel = ch) <- img
           })
         }
       }
@@ -1242,7 +1242,7 @@ writeZarrArrayInImage <- function(
     zarrcreateGroup(zarr_path, paste0(name, "/spat_", spat))
 
     # write coordinates
-    coords <- vrCoordinates(object, spatial_name = spat)
+    coords <- vrCoordinates(object, spatial = spat)
     if (!inherits(coords, c("DelayedArray", "IterableMatrix")) || replace) {
       if (verbose) {
         message("Writing '", name, "' coordinates")
@@ -1262,15 +1262,15 @@ writeZarrArrayInImage <- function(
       # Rarr::ZarrArray doesnt have rownames
       rownames(coords) <- vrSpatialPoints(object)
 
-      vrCoordinates(object, spatial_name = spat) <- coords
+      vrCoordinates(object, spatial = spat) <- coords
     }
 
     # for each channel
-    channels <- vrImageChannelNames(object, name = spat)
+    channels <- vrImageChannelNames(object, spatial = spat)
     if (!all(grepl("No Channels", channels))) {
       for (ch in channels) {
         # get image and write to h5
-        img <- vrImages(object, name = spat, channel = ch, as.raster = TRUE)
+        img <- vrImages(object, spatial = spat, channel = ch, as.raster = TRUE)
 
         # write image
         if (!inherits(img, "ImageArray") || replace) {
@@ -1296,7 +1296,7 @@ writeZarrArrayInImage <- function(
             verbose = FALSE
           )
           suppressWarnings({
-            vrImages(object, name = spat, channel = ch) <- img
+            vrImages(object, spatial = spat, channel = ch) <- img
           })
         }
       }
@@ -1415,12 +1415,12 @@ writeZarrArrayInImage <- function(
   spatial_names <- vrSpatialNames(object)
   for (spat in spatial_names) {
     # for each channel
-    channels <- vrImageChannelNames(object, name = spat)
+    channels <- vrImageChannelNames(object, spatial = spat)
     for (ch in channels) {
       cur_path <- try(
         DelayedArray::path(vrImages(
           object,
-          name = spat,
+          spatial = spat,
           channel = ch,
           as.raster = TRUE
         )),
@@ -1552,22 +1552,22 @@ setMethod(
   for (spat in spatial_names) {
     # coordinates
     object@image[[spat]]@coords <-
-      .modify_seeds(vrCoordinates(object, spatial_name = spat), function(x) {
+      .modify_seeds(vrCoordinates(object, spatial = spat), function(x) {
         .shorten_assay_links_data(x)
       })
 
     # for each channel
-    channels <- vrImageChannelNames(object, name = spat)
+    channels <- vrImageChannelNames(object, spatial = spat)
     if (!all(grepl("No Channels", channels))) {
       for (ch in channels) {
-        img <- vrImages(object, name = spat, channel = ch, as.raster = TRUE)
+        img <- vrImages(object, spatial = spat, channel = ch, as.raster = TRUE)
         for (i in seq_len(length(img@levels))) {
           img[[i]] <- .modify_seeds(img[[i]], function(x) {
             .shorten_assay_links_data(x)
           })
         }
         suppressWarnings({
-          vrImages(object, name = spat, channel = ch) <- img
+          vrImages(object, spatial = spat, channel = ch) <- img
         })
       }
     }
@@ -1714,15 +1714,15 @@ setMethod(
   for (spat in spatial_names) {
     # coordinates
     object@image[[spat]]@coords <-
-      .modify_seeds(vrCoordinates(object, spatial_name = spat), function(x) {
+      .modify_seeds(vrCoordinates(object, spatial = spat), function(x) {
         .restore_absolute_links(x, dir)
       })
 
     # for each channel
-    channels <- vrImageChannelNames(object, name = spat)
+    channels <- vrImageChannelNames(object, spatial = spat)
     if (!all(grepl("No Channels", channels))) {
       for (ch in channels) {
-        img <- vrImages(object, name = spat, channel = ch, as.raster = TRUE)
+        img <- vrImages(object, spatial = spat, channel = ch, as.raster = TRUE)
         # TODO: for now length method for ImageArray
         for (i in seq_len(length(img@levels))) {
           img[[i]] <- .modify_seeds(img[[i]], function(x) {
@@ -1730,7 +1730,7 @@ setMethod(
           })
         }
         suppressWarnings({
-          vrImages(object, name = spat, channel = ch) <- img
+          vrImages(object, spatial = spat, channel = ch) <- img
         })
       }
     }